Two changes to team color handling.
Extract the flag object to its own prefab. Revise how we update the selection color (so that it works for the first player): Add a "None" value to be the initial value, so our hook is called when it is first set. Otherwise, Teal changes to Teal for first player, and the value is not synced, so hook is not triggered.

Revise the outline configuration, to prevent a flicker from default colour. Initially (in prefab) set the outline to be hidden, but enabled. When a unit is spawned, change the colour, show the outline, and disable it all at once. Also remove the unity event for selection changed; it only added complexity.